Revert "moved test to same folder as component"
This reverts commit 7e61bf4bbf11dc3770854f1c10c13c50f402a3c2.
